Add gum theme kdl with template and migration, --if-empty flag for tmp-clear, fix typo FEATHERSTL name

+++ script/helpers/tmp-clear.sh
@@ -6,4 +6,6 @@ if [[ -z "$FEATHERT" ]]; then
 	echo "FEATHERT is not set!"
 	exit 1
 fi
-rm -rf "$FEATHERT"
+if [[ "--if-empty" != "$1" ]] || ! [ "$(ls -A "$FEATHERT")" ]; then
+	rm -rf "$FEATHERT"
+fi
